refactor(mobile): split email pipeline
