+++ apps/mobile/features/email-capture/services/email-pipeline-service/types.ts
@@ -1,0 +1,50 @@
+export type {
+  CaptureEvidenceRowsInput,
+  CaptureEvidenceSaveInput,
+  DefaultAccountInput,
+  DuplicateLookupOutcome,
+  DuplicateProcessedEmailRowInput,
+  EmailBatchContext,
+  EmailBatchPlan,
+  EmailMetric,
+  EmailQueue,
+  EmailTransactionContext,
+  IncomingEmailPersistenceInput,
+  IncomingParseOutcome,
+  LinkCaptureEvidenceInput,
+  MerchantRuleEffectInput,
+  PersistedTransactionContext,
+  PipelineRuntime,
+  ProcessEmailsInput,
+  ProcessedEmailStatusEffectInput,
+  ProcessRetriesInput,
+  RetryBatchContext,
+  RetryDuplicateOutcome,
+  RetryParseOutcome,
+  RetryScheduleEffectInput,
+  RetrySuccessEffectInput,
+  RetryTransactionContext,
+  SaveRetryTransactionInput,
+  SaveTransactionInput,
+  TrackSavedTransactionInput,
+  UnparsedIncomingEmailKind,
+  UnparsedProcessedEmailRowInput,
+} from "./internal-types";
+export type {
+  CategoryId,
+  CreateEmailPipelineServiceDeps,
+  EmailPipelineService,
+  EmailSaveStatus,
+  IsoDateTime,
+  LlmParsedTransaction,
+  PipelineResult,
+  ProcessEmails,
+  ProcessedEmailId,
+  ProcessedEmailRow,
+  ProcessRetries,
+  ProgressCallback,
+  RawEmail,
+  RetryResult,
+  TransactionId,
+  UserId,
+} from "./public-types";
